Moved articles endpoints
